+++ odoo_dynamic_dashboard/__init__.py
@@ -19,5 +19,6 @@
 #    If not, see <http://www.gnu.org/licenses/>.
 #
 #############################################################################
-from . import models
 from . import controllers
+from . import models
+from . import wizard