+++ odoo_dynamic_dashboard/models/__init__.py
@@ -20,6 +20,6 @@
 #
 #############################################################################
 from . import dashboard_block
-from . import dashboard_block_line
 from . import dashboard_menu
+from . import dashboard_theme
 from . import domain_to_sql