+++ odoo_dynamic_dashboard/models/dashboard_menu.py
@@ -23,77 +23,58 @@ from odoo import api, fields, models
 
 
 class DashboardMenu(models.Model):
-    """
-    This is the class DashboardMenu which is the subclass of the class Model
-    which is here used to create the model dashboard.menu.
-    """
+    """Class to create new dashboard menu"""
     _name = "dashboard.menu"
     _description = "Dashboard Menu"
 
-    name = fields.Char(string="Name", help="Name of the dashboard")
-    parent_id = fields.Many2one('ir.ui.menu', string="Menu",
-                                help="Parent of the dashboard")
+    name = fields.Char(string="Name", ondelete='cascade',
+                       help="Enter a name for the dashboard menu")
+    menu_id = fields.Many2one('ir.ui.menu', string="Parent Menu",
+                              help="Parent Menu Location of New Dashboard",
+                              ondelete='cascade')
     group_ids = fields.Many2many('res.groups', string='Groups',
-                                 related='parent_id.groups_id',
-                                 help="User need to be at least in one "
-                                      "of these groups to see the menu")
+                                 related='menu_id.groups_id',
+                                 help="User need to be at least in one of "
+                                      "these groups to see the menu")
     client_action_id = fields.Many2one('ir.actions.client',
-                                       string='Client Action',
-                                       help="Client Action Related "
-                                            "to the dashboard")
-    menu_id = fields.Many2one('ir.ui.menu', string="Created Menu",
-                              help="Created menu")
+                                       string="Client Action",
+                                       help="Client action of the "
+                                            "corresponding dashboard menu")
 
     @api.model
     def create(self, vals):
-        """
-        Summary:
-            This is the create function of the model DashboardMenu which is
-            triggered when creating a new record in this model.
-        Args:
-            vals:
-                The values when the user creating a new record.
-        Returns:
-            res:
-                Returns the created record at the end
-        """
-        values = {
+        """Function to create new dashboard menu"""
+        action_id = self.env['ir.actions.client'].create({
             'name': vals['name'],
-            'tag': 'owl.dynamic_dashboard',
-        }
-        action_id = self.env['ir.actions.client'].create(values)
-        vals['client_action_id'] = action_id.id
-        menu_id = self.env['ir.ui.menu'].create({
-            'name': vals['name'],
-            'parent_id': vals['parent_id'],
-            'action': f'ir.actions.client,{action_id.id}'
+            'tag': 'OdooDynamicDashboard',
         })
-        res = super(DashboardMenu, self).create(vals)
-        res.menu_id = menu_id.id
-        return res
+        vals['client_action_id'] = action_id.id
+        self.env['ir.ui.menu'].create({
+            'name': vals['name'],
+            'parent_id': vals['menu_id'],
+            'action': 'ir.actions.client,%d' % (action_id.id,)
+        })
+        return super(DashboardMenu, self).create(vals)
 
     def write(self, vals):
-        """
-        Summary:
-            This is the write function of the model DashboardMenu which is
-            triggered when changing any value in the corresponding record.
-        Args:
-            vals:
-                The values when the user creating a editing an record.
-        Returns:
-            Returns the updated record at the end
-        """
-        if self.menu_id:
-            self.menu_id.update(vals)
+        """Function to save edited data in dashboard menu"""
+        for rec in self:
+            client_act_id = rec['client_action_id'].id
+            self.env['ir.ui.menu'].search(
+                [('parent_id', '=', rec['menu_id'].id),
+                 ('action', '=', f'ir.actions.client,{client_act_id}')]).write({
+                'name': vals['name'] if 'name' in vals.keys() else rec['name'],
+                'parent_id': vals['menu_id'] if 'menu_id' in vals.keys() else
+                rec['menu_id'],
+                'action': f'ir.actions.client,{client_act_id}'
+            })
         return super(DashboardMenu, self).write(vals)
 
     def unlink(self):
-        """
-        Summary:
-            This is the unlink function of the model DashboardMenu which is
-            triggered when unlinking any record in this model.
-        Returns:
-            Returns the record delete.
-        """
-        self.menu_id.unlink()
+        """Delete dashboard along with menu item"""
+        for rec in self:
+            self.env['ir.ui.menu'].search(
+                [('parent_id', '=', rec['menu_id'].id),
+                 ('action', '=',
+                  f'ir.actions.client,{rec["client_action_id"].id}')]).unlink()
         return super(DashboardMenu, self).unlink()