+++ odoo_dynamic_dashboard/doc/RELEASE_NOTES.md
@@ -1,7 +1,13 @@
 ## Module <odoo_dynamic_dashboard>
 
-#### 02.03.2024
+#### 18.05.2024
 #### Version 17.0.1.0.0
-#### ADD
+##### ADD
+- Initial commit for Odoo Dynamic Dashboard
 
-- Initial Commit for Odoo Dynamic Dashboard
+## Module <odoo_dynamic_dashboard>
+
+#### 20.07.2024
+#### Version 17.0.2.0.0
+##### UPDT
+- New Features has been Added.